Removed some unnecessary GUI files, updated UtilityCart & Shopping Spree
In the process of removing a lot of garbage from the GUI pack. Moved the utility cart "grab point" so it no longer clips through the floor & unchecked "use gravity" on the rigidbody. Added UI arrows to the initial utility carts on Shopping Spree in order to direct the player to them